Let the phone plan excess, and show the reason it was given
The job form refused any quantity over what the order had left —
"Qty for X exceeds pending Y" — and refused it locally, before a
request was ever sent. The server stopped working that way: a line may
be planned up to 20% over the ORDERED quantity with no comment, and
further with a written reason, drawing the extra yarn from stock as it
goes (services/excessPlanning.js).

So the app was telling the floor it could not do something the system
permits, with no way to find out otherwise, while the same planner
could do it at a desk. The rule here now mirrors the server's:

  · every ordered line is offered, not only the ones with something
    left — a fully-planned line can still take excess, and hiding it
    made that impossible from here;
  · excess is measured against the ordered quantity, not what remains,
    so two jobs each inside the allowance cannot add up to 40% over;
  · past 20% a reason is asked for and held to the same 8-character
    floor the server enforces;
  · the yarn warning is shown before the button is pressed rather than
    arriving as a 409 afterwards.

The two 409s the server can still return are now told apart. One wants
a sentence typed and the other wants yarn bought, and "Failed to create
job order" told the person on the floor neither; the stock refusal now
names the material and the shortfall.

The order detail grows the excess card the web has: which job planned
over, by how much, the yarn it drew, and the reason — or a note that
the line was inside the allowance and nobody was asked, because a blank
there reads as withheld.

Fixed on the way past: ElasticQtyModel.quantity was declared int while
its factory assigned .toDouble() to it. That type-checks only because
the JSON value arrives as dynamic, so the mismatch waited to surface at
runtime, as a cast error building the order, the first time a quantity
was not a whole number of metres. Metres are fractional — the order
form takes two decimals — so it is a double now.

Also added /order-pnl and /samples to the feature catalogue. Without
them an admin working from the phone cannot grant either feature to
anybody.

+++ src/features/Job/models/ElasticQtyModel.dart
@@ -1,7 +1,13 @@
 class ElasticQtyModel {
   final String elasticId;
   final String elasticName;
-  final int quantity;
+
+  // Metres, and metres are fractional — the order form takes two
+  // decimals. Declared `int` while the factory below assigned
+  // `.toDouble()` to it, which type-checks only because the JSON value
+  // arrives as `dynamic`: the mismatch surfaced at RUNTIME, as a cast
+  // error building the order, the first time a quantity was not whole.
+  final double quantity;
 
   ElasticQtyModel({
     required this.elasticId,
